+++ aiida/orm/data/__init__.py
@@ -8,7 +8,6 @@
 # For further information please visit http://www.aiida.net               #
 ###########################################################################
 from abc import ABCMeta
-from collections import MutableSequence
 
 try:
     from functools import singledispatch
@@ -18,7 +17,6 @@ except ImportError:
 from aiida.orm.node import Node
 from aiida.common.links import LinkType
 from aiida.common.lang import override
-from aiida.common.exceptions import ModificationNotAllowed
 
 
 @singledispatch
@@ -462,109 +460,3 @@ class BaseType(Data):
                 "When specifying dbnode it can be the only kwarg"
 
         return kwargs
-
-
-
-class List(Data, MutableSequence):
-    """
-    Class to store python lists as AiiDA nodes
-    """
-    _LIST_KEY = 'list'
-
-    def __init__(self):
-        super(List, self).__init__()
-        self._set_list(list())
-
-    def __getitem__(self, item):
-        return self._get_list()[item]
-
-    def __setitem__(self, key, value):
-        l = self._get_list()
-        l[key] = value
-        if not self._using_list_reference():
-            self._set_list(l)
-
-    def __delitem__(self, key):
-        l = self._get_list()
-        del l[key]
-        if not self._using_list_reference():
-            self._set_list(l)
-
-    def __len__(self):
-        return len(self._get_list())
-
-    def __str__(self):
-        return self._get_list().__str__()
-
-    def append(self, value):
-        l = self._get_list()
-        l.append(value)
-        if not self._using_list_reference():
-            self._set_list(l)
-
-    def extend(self, L):
-        l = self._get_list()
-        l.extend(L)
-        if not self._using_list_reference():
-            self._set_list(l)
-
-    def insert(self, i, value):
-        l = self._get_list()
-        l.insert(i, value)
-        if not self._using_list_reference():
-            self._set_list(l)
-
-    def remove(self, value):
-        del self[value]
-
-    def pop(self, **kwargs):
-        l = self._get_list()
-        l.pop(**kwargs)
-        if not self._using_list_reference():
-            self._set_list(l)
-
-    def index(self, value):
-        return self._get_list().index(value)
-
-    def count(self, value):
-        return self._get_list().count(value)
-
-    def sort(self, cmp=None, key=None, reverse=False):
-        l = self._get_list()
-        l.sort(cmp, key, reverse)
-        if not self._using_list_reference():
-            self._set_list(l)
-
-    def reverse(self):
-        l = self._get_list()
-        l.reverse()
-        if not self._using_list_reference():
-            self._set_list(l)
-
-    def _get_list(self):
-        try:
-            return self.get_attr(self._LIST_KEY)
-        except AttributeError:
-            self._set_list(list())
-            return self.get_attr(self._LIST_KEY)
-
-    def _set_list(self, list_):
-        if not isinstance(list_, list):
-            raise TypeError("Must supply list type")
-        self._set_attr(self._LIST_KEY, list_)
-
-    def _using_list_reference(self):
-        """
-        This function tells the class if we are using a list reference.  This
-        means that calls to self.get_list return a reference rather than a copy
-        of the underlying list and therefore self._set_list need not be called.
-        This knwoledge is essential to make sure this class is performant.
-
-        Currently the implementation assumes that if the node needs to be
-        stored then it is using the attributes cache which is a reference.
-
-        :return: True if using self._get_list returns a reference to the
-            underlying sequence.  False otherwise.
-        :rtype: bool
-        """
-        return not self.is_stored
